backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "django-ninja>=1.4.3",
    "google-genai>=1.36.0",
    "ninja-schema>=0.14.3",
    "pydantic-to-typescript2>=1.0.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
